Organized XMLs and added templates/custom/base directory. Updated base directories.
All XMLs in templates and in templates/custom are now organized into subdirectories typeLinked, strictClock, or UCLD.  BE SURE TO UPDATE YOUR SCRIPTS (ken_templates wasn't changed).

Added directory templates/custom/base that now has base templates. We plan to generate the other templates from these to ease the burden on editing and maintaining long term.

All templates in base were updated.  Unneeded clockPriors (unused variable) and nonZeroRatePriors (prior on rate indicators) were removed. rateIndicators are now defined at the beginning of XMLs for clarity and are no longer stateNodes (now if someone tries to operate on rateIndicators they'll get an error). clockRate (unused) is no longer defined as a stateNode and is only defined in the branchRateModel (definition in stateNodes was redundant).